refactored speech components
