+++ src/content/components/FeedbackForm.js
@@ -1,14 +1,12 @@
 import React, { useState, useEffect } from 'react';
 import jwt_decode from 'jwt-decode';
 import FlashMessage from '../components/FlashMessage';
-import SpeechRecognition, { useSpeechRecognition } from 'react-speech-recognition';
 import Button from '@material-ui/core/Button';
-import TextareaAutosize from '@material-ui/core/TextareaAutosize';
 import Grid from '@material-ui/core/Grid';
 import axios from 'axios';
 import ListeningButtons from './ListeningButtons';
 import WriteForm from './WriteForm';
-import SpeechForm from './SpeechForm';
+import GetSpeech from './GetSpeech';
 
 const FeedbackForm = (props) => {
     const [isListening, setIsListening] = useState(false)
@@ -37,37 +35,6 @@ const FeedbackForm = (props) => {
         setInputs({...inputs, [e.target.name]: e.target.value})
     }
     
-    const startListening = () => {
-        SpeechRecognition.startListening({ continuous: true })
-        setIsListening(true)
-    }
-
-    const stopListening = () => {
-        SpeechRecognition.stopListening()
-        setIsListening(false)
-    }
-    
-    const {interimTranscript, transcript, finalTranscript, resetTranscript } = useSpeechRecognition()
-
-    // for speech form only
-    useEffect(() => {
-        if (interimTranscript !== '') {
-            // console.log('Got interim result:', interimTranscript)
-            setInputs({
-                answer: transcript,
-                content: props.selectedQuestion,
-                category: props.selectedCategory
-            })
-        }
-        if (finalTranscript !== '') {
-            // console.log('Got final result:', finalTranscript)
-        }
-    }, [interimTranscript, finalTranscript]);
-
-    if (!SpeechRecognition.browserSupportsSpeechRecognition()) {
-        return null
-    }
-
     const handleSubmit = e => {
       e.preventDefault();
       // mount props on input without setting state to handle edgecase
@@ -91,31 +58,17 @@ const FeedbackForm = (props) => {
     }
 
     const displaySpeechForm = (
-        <Grid container spacing={12}>
-            <Grid item xs={12} className="feedback-buttons-row">
-                <ListeningButtons 
-                    startListening={startListening}
-                    stopListening={stopListening}
-                    resetTranscript={resetTranscript}
-                />
-            </Grid>
-            <Grid item xs={6}>
-                <SpeechForm
-                handleSubmit={handleSubmit}
-                transcript={transcript}
-                />
-            </Grid>
-        </Grid>        
+        <GetSpeech 
+        setIsListening={setIsListening}
+        setInputs={setInputs}
+        handleSubmit={handleSubmit}
+        /> 
     )
 
     const displayWriteForm = (
         <Grid container spacing={6}>
             <Grid item xs={12} className="feedback-buttons-row">
-                <ListeningButtons
-                 startListening={startListening}
-                 stopListening={stopListening}
-                 resetTranscript={resetTranscript}
-                />
+                <ListeningButtons />
             </Grid>
             <Grid item xs={6}>
                 <WriteForm
